Improved code
With my code before my eigenvaules were diverging :O But with this code they are not. Also it features a variable called shiftedre which will allows me to offset my re. 
The function lsf_quick in LSF_hmwk5 allows the user to quickly do a polynomial fit (only for a power of 2 though). This was useful for determining if my values were diverging or converging.